Updated participant business logic

+++ src/main/java/com/tournabay/api/service/ParticipantService.java
@@ -2,10 +2,8 @@ package com.tournabay.api.service;
 
 import com.tournabay.api.exception.BadRequestException;
 import com.tournabay.api.exception.ResourceNotFoundException;
-import com.tournabay.api.model.Participant;
-import com.tournabay.api.model.ParticipantStatus;
-import com.tournabay.api.model.Tournament;
-import com.tournabay.api.model.User;
+import com.tournabay.api.model.*;
+import com.tournabay.api.payload.UpdateParticipantRequest;
 import com.tournabay.api.repository.ParticipantRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,6 +16,7 @@ import java.util.List;
 public class ParticipantService {
     private final ParticipantRepository participantRepository;
     private final UserService userService;
+    private final TeamService teamService;
 
     public Participant save(Participant participant) {
         return participantRepository.save(participant);
@@ -41,7 +40,7 @@ public class ParticipantService {
     }
 
     public List<Participant> setParticipantsStatus(List<Participant> participants, ParticipantStatus status) {
-        participants.forEach(participant -> participant.setParticipantStatus(status));
+        participants.forEach(participant -> participant.setStatus(status));
         return participantRepository.saveAll(participants);
     }
 
@@ -57,7 +56,8 @@ public class ParticipantService {
         return participant;
     }
 
-    public void deleteById(Long participantId) {
+    public void deleteById(Long participantId, Tournament tournament) {
+        if (!tournament.containsParticipantById(participantId)) throw new BadRequestException("Participant doesn't exist in tournament");
         participantRepository.deleteById(participantId);
     }
 
@@ -65,8 +65,33 @@ public class ParticipantService {
         User user = userService.addUserByOsuId(osuId);
         return Participant.builder()
                 .user(user)
+                .discordId(user.getDiscordId())
                 .joinedAt(LocalDateTime.now())
-                .participantStatus(ParticipantStatus.ACCEPTED)
+                .status(ParticipantStatus.ACCEPTED)
                 .build();
+    }
+
+    public Participant updateParticipant(Participant participant, UpdateParticipantRequest body, Tournament tournament) {
+        if (tournament instanceof TeamBasedTournament teamBasedTournament) {
+            // TODO: Limit team size when settings are done for TeamBasedTournaments
+            // Detach the participant from the team if he's in any
+            teamBasedTournament.getTeams()
+                    .stream()
+                    .filter(team -> team.getParticipants().contains(participant))
+                    .findFirst()
+                    .ifPresent(team -> {
+                        team.getParticipants().remove(participant);
+                        teamService.save(team);
+                    });
+            Team team = teamService.findById(body.getTeamId());
+            team.getParticipants().add(participant);
+            teamService.save(team);
+            participant.setDiscordId(body.getDiscordId());
+            return participantRepository.save(participant);
+        } else if (tournament instanceof PlayerBasedTournament playerBasedTournament) {
+            participant.setDiscordId(body.getDiscordId());
+            return participantRepository.save(participant);
+        }
+        throw new BadRequestException("Tournament type not supported");
     }
 }
